bspctl/doctor: add --psi-calibrate flag for PSI threshold tuning
There was no workflow for users to determine appropriate
pressure_max_cpu/io/memory values before committing them to their
config. Calibration values guessed without sampling the idle-system
baseline are likely to cause unnecessary build pauses on lightly
loaded hosts or no throttling at all on saturated ones.

Add --psi-calibrate/-C to `bspctl doctor`. It samples avg10 PSI
readings every 0.5 s until Ctrl+C, then prints recommendations at
peak + 20% headroom (clamped to 95, memory floor 20). When PSI files
are unreadable the subcommand exits 0 with a clear message instead of
crashing. Tests cover the recommendation math and the unavailable-PSI
CLI path.

+++ src/bspctl/config.py
@@ -99,9 +99,9 @@ class BuildConfig:
     sstate_dir: str | None = field(default=None)
     sstate_mirrors: str | None = field(default=None)
     scheduler: str | None = field(default=None)
-    pressure_max_cpu: int | None = field(default=None)
-    pressure_max_io: int | None = field(default=None)
-    pressure_max_memory: int | None = field(default=None)
+    pressure_max_cpu: float | None = field(default=None)
+    pressure_max_io: float | None = field(default=None)
+    pressure_max_memory: float | None = field(default=None)
 
     @property
     def workspace_subdir(self) -> str:
